requirements.txt
annotated-types==0.7.0
cffi==1.17.1
colorama==0.4.6
cryptography==44.0.0
cx-Oracle==8.3.0
et_xmlfile==2.0.0
numpy==2.2.2
openpyxl==3.1.5
oracledb==2.5.1
pandas==2.2.3
psycopg2==2.9.10
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
six==1.17.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.1
